+++ src/tipos-servicio/tipos-servicio.controller.ts
@@ -1,6 +1,7 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Query, UseGuards, UsePipes, ValidationPipe } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, Post, Put, Query, UnauthorizedException, UseGuards, Headers, UsePipes, ValidationPipe } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 import { ApiBearerAuth } from '@nestjs/swagger';
+import { AuthService } from 'src/auth/auth.service';
 import { TiposServicioDto } from './tipos-servicio.dto';
 import { TiposServicio } from './tipos-servicio.entity';
 import { TiposServicioService } from './tipos-servicio.service';
@@ -8,7 +9,7 @@ import { TiposServicioService } from './tipos-servicio.service';
 @Controller('tipos_servicio')
 export class TiposServicioController{
 
-constructor(readonly service:TiposServicioService){}
+constructor(readonly service:TiposServicioService,private authServ:AuthService){}
 
 
 
@@ -20,6 +21,8 @@ async getCount(@Query() query:any){
 }
 
 
+@ApiBearerAuth()
+@UseGuards(AuthGuard('jwt'))
 @Get('existTipoByDescripcion/:desc')
 async existTipoByDescripcion(@Param('desc') desc:string) {
     if(await this.service.getTipoByDescripcion(desc)){
@@ -48,7 +51,10 @@ async getOne(@Param('id') id:number){
 @ApiBearerAuth()
 @UseGuards(AuthGuard('jwt'))
 @Post()
-async createOne(@Body() dto:TiposServicioDto){
+async createOne(@Headers() headers:any,@Body() dto:TiposServicioDto){
+    if(!(await this.authServ.isTokenAdmin(headers))){
+        throw new UnauthorizedException('Necesitas ser administrador!');
+    }
     return await this.service.createOne(dto);
 }
 
@@ -56,7 +62,10 @@ async createOne(@Body() dto:TiposServicioDto){
 @ApiBearerAuth()
 @UseGuards(AuthGuard('jwt'))
 @Put(':id')
-async editOne(@Param('id') id:number,@Body() dto:TiposServicioDto){
+async editOne(@Headers() headers:any,@Param('id') id:number,@Body() dto:TiposServicioDto){
+    if(!(await this.authServ.isTokenAdmin(headers))){
+        throw new UnauthorizedException('Necesitas ser administrador!');
+    }
     return await this.service.editOne(id,dto);
 }
 
@@ -64,7 +73,10 @@ async editOne(@Param('id') id:number,@Body() dto:TiposServicioDto){
 @ApiBearerAuth()
 @UseGuards(AuthGuard('jwt'))
 @Delete(':id')
-async deleteOne(@Param('id') id:number){
+async deleteOne(@Headers() headers:any,@Param('id') id:number){
+    if(!(await this.authServ.isTokenAdmin(headers))){
+        throw new UnauthorizedException('Necesitas ser administrador!');
+    }
     return await this.service.deleteOne(id); 
 }
